requirements.txt
openai
python-dotenv
dateparser
datetime